+++ core-ng/src/main/java/core/framework/internal/redis/Protocol.java
@@ -63,6 +63,8 @@ final class Protocol {
     }
 
     static class Command {
+        static final byte[] AUTH = Strings.bytes("AUTH");
+
         static final byte[] GET = Strings.bytes("GET");
         static final byte[] SET = Strings.bytes("SET");
         static final byte[] PEXPIRE = Strings.bytes("PEXPIRE");
@@ -104,8 +106,6 @@ final class Protocol {
         static final byte[] ZRANGE = Strings.bytes("ZRANGE");
         static final byte[] ZRANGEBYSCORE = Strings.bytes("ZRANGEBYSCORE");
         static final byte[] ZREM = Strings.bytes("ZREM");
-
-        static final byte[] AUTH = Strings.bytes("AUTH");
     }
 
     static class Keyword {